Library Reference

AlgebraicABMs.Distributions.weibullpar — Method

Get the shape and scale parameters for a Weibull distribution with a given mean and variance. This code is a translation of the R "mixdist" package's function

AlgebraicABMs.ABMs.ABM — Type

An agent-based model.

AlgebraicABMs.ABMs.ABMFlow — Type

Continuous dynamics

AlgebraicABMs.ABMs.ABMRule — Type

A stochastic rewrite rule with a dependent hazard rate

AlgebraicABMs.ABMs.AbsDynamics — Type

A type which implements AbsDynamics must be able to compiled to an ODE for some set of variables.

AlgebraicABMs.ABMs.AbsHomSet — Type

A collection of timers associated at runtime w/ an ABMRule

AlgebraicABMs.ABMs.AbsTimer — Type

Something that can produce a ACSetTransformation × clocktime → hazard_rate

AlgebraicABMs.ABMs.ClosureState — Type

A closure which accepts a match morphism and returns a hazard_rate. This is a timer which cannot depend on the absolute clock time.

AlgebraicABMs.ABMs.ClosureTime — Type

A closure which accepts a clocktime and returns a hazard_rate. This is a timer which cannot depend on the match data nor ACSet state.

AlgebraicABMs.ABMs.EmptyP — Type

Empty patterns have (one) trivial pattern match

AlgebraicABMs.ABMs.FullClosure — Type

A closure which accepts a ACSetTransformation and returns a function of type clocktime → hazard_rate

AlgebraicABMs.ABMs.RawODE — Type

Use raw Julia functions to define an ODE

AlgebraicABMs.ABMs.RegularP — Type

Default case, where pattern matches should be found via (incremental) homomorphism search and represented explicitly, each with own events getting scheduled.

AlgebraicABMs.ABMs.RepresentableP — Type

A pattern match from a coproduct of representables is just a choice of parts in the codomain. E.g. matching L = •→• • • is just a random choice of edge and two random vertices.

The vector of ints refers to parts of L which are the counits of the left kan extensions that define the representables (usually this is just wherever the colimit leg sends 1, as there is often just one X part in the representable X).

WARNING: this is only viable if the timer associated with the rewrite rule is symmteric with respect to the disjoint representables and has a simple exponential timer.

AlgebraicABMs.ABMs.RuntimeABM — Type

Data @structhashequal structure for maintaining simulation information while running an ABM

AlgebraicABMs.ABMs.Traj — Type

A trajectory of an ABM: each event time and result of save.

AlgebraicABMs.ABMs.init_homset — Method

Initialize runtime hom-set given the rule and the initial state

AlgebraicABMs.ABMs.is_exp — Method

Check if a hazard rate is a simple exponential

AlgebraicABMs.ABMs.mk_prob — Method

Construct an ODE for a given ACSet state. Return a mapping which allows to go from index to AttrType+index.

AlgebraicABMs.ABMs.pattern_type — Method

Analyze a pattern to find the most efficient pattern type for it.

Because ACSet types do not know their own equations, we may have to pass the schema as an argument in order to compute representables that would otherwise be infinite.

Even if the pattern is a coproduct of representables, we cannot use the efficient encoding unless the distribution is either an exponential (or a single dirac delta - not yet supported).

AlgebraicABMs.ABMs.run! — Method

Run an ABM, creating a fresh runtime + trajectory.

save - function applied to the ACSet state to produce the data that gets stored for every change in the model dt - timestep for checking discrete events when running ODE dynamics.

AlgebraicABMs.ABMs.validate — Method

Check that RuntimeABM incremental hom sets have all valid homs.

AlgebraicABMs.Upstream.pops! — Method

Pop the next random event, advance the clock

AlgebraicRewriting.Rewrite.Utils.get_match — Method

Get match returns a randomly chosen morphism for the aggregate rule